Framework/agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Pablo Ruiz", email = "p_ruizruiz@apple.com" }
]
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.22",
    "lmstudio>=1.2.0",
    "ollama>=0.4.7",
    "openai>=1.69.0",
    "pillow>=11.1.0",
    "pydantic-settings>=2.8.1",
    "pydantic>=2.11.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "vertexai>=1.71.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "fastapi>=0.115.12",
    "pytest>=8.3.5",
    "ruff>=0.11.2",
    "streamlit>=1.44.0",
]
